fix: unblock wasm32 aero-wasm check
- Remove duplicate AeroGpuPciConfigDevice definitions in aero-machine

- Fix duplicate BlockLowerer field initializer in aero-jit-x86

- Address clippy -D warnings failures in aero-jit-x86 tier2 + tests

+++ crates/aero-jit-x86/src/tier2/wasm_codegen.rs
@@ -820,7 +820,7 @@ impl Emitter<'_> {
                                 //
                                 // This is equivalent under wrapping arithmetic and tends to
                                 // generate better code in Wasm engines.
-                                let sh = u32::from(scale.trailing_zeros()) as i64;
+                                let sh = scale.trailing_zeros() as i64;
                                 self.f.instruction(&Instruction::I64Const(sh));
                                 self.f.instruction(&Instruction::I64Shl);
                             } else {
@@ -833,7 +833,7 @@ impl Emitter<'_> {
                         self.emit_operand(index);
                         if scale != 1 {
                             if scale.is_power_of_two() {
-                                let sh = u32::from(scale.trailing_zeros()) as i64;
+                                let sh = scale.trailing_zeros() as i64;
                                 self.f.instruction(&Instruction::I64Const(sh));
                                 self.f.instruction(&Instruction::I64Shl);
                             } else {
